Implement image analysis and brew management API endpoints
